feat: add Anywhere Cache samples and integration tests
Implement Go SDK samples for Anywhere Cache features in storage/control,
including create, get, list, update, pause, resume, and disable.
Add integration tests for the Anywhere Cache lifecycle in anywhere_cache_test.go.

Target API:
- CreateAnywhereCache (LRO)
- GetAnywhereCache
- ListAnywhereCaches (Iterator)
- UpdateAnywhereCache (LRO)
- PauseAnywhereCache
- ResumeAnywhereCache
- DisableAnywhereCache
- *controlpb.AnywhereCache

The implementation follows the requested 2026 copyright, region tags,
LRO handling with .Wait(ctx), and specific commentary.
Tests are located in their own file `anywhere_cache_test.go`.
